+++ demo/src/components/Samples.vue
@@ -51,6 +51,8 @@ export default {
 
       manualStringValue: '8:15 pm',
 
+      customCloseBtnValue: '10:05',
+
       autoScrollData1: '08:40',
       autoScrollData2: '5:30:20 pm',
 
@@ -80,7 +82,10 @@ export default {
         { title: 'Customized Picker Labels', anchor: 'customPickerLabels' },
         { title: 'Adjust Input Width', anchor: 'inputWidth' },
         { title: 'Auto-Scroll', anchor: 'autoScroll' },
-        { title: 'More Powerful format String', anchor: 'morePowerfulFormat' }
+        { title: 'More Powerful format String', anchor: 'morePowerfulFormat' },
+        { title: 'Customized Buttons And Icon', anchor: 'customButtonIcon' },
+        { title: 'Fixed Dropdown Button', anchor: 'fixedDropdownButton' },
+        { title: 'Drop Direction', anchor: 'dropDirection' }
       ]
     }
   },
@@ -145,7 +150,7 @@ section#mostlyUsedSamples
     template(v-slot:codes)
       highlight-code(lang="html" data-title="HTML")
         pre
-          | &lt;!-- 12-hour sampleaa 1 --&gt;
+          | &lt;!-- 12-hour sample 1 --&gt;
           | &lt;vue-timepicker format="hh:mm A"&gt;&lt;/vue-timepicker&gt;
           | &nbsp;
           | &lt;!-- 12-hour sample 2 --&gt;
@@ -461,6 +466,7 @@ section#mostlyUsedSamples
         | &lt;vue-timepicker hide-clear-button&gt;&lt;/vue-timepicker&gt;
     template(v-slot:preview)
       vue-timepicker(hide-clear-button)
+
 
   //- Disable Picker
   sample-block#disablePicker
@@ -851,6 +857,143 @@ section#mostlyUsedSamples
         | &nbsp;
         vue-timepicker(format="a" input-width="60px")
 
+  //- Custom Button And Icon
+  sample-block#customButtonIcon
+    template(v-slot:title) Customized Buttons And Input Icon
+    template(v-slot:description)
+      p You can customize the clear button, and the dropdown button with your own icon/image starts from <code>v1.1.4+</code>. There's an additional slot for the input icon as well.
+      p Please note that Vue v2.6.0+ introduces a significant update of the <b>Named Slots</b> syntax. Check the <a href="https://vuejs.org/v2/guide/components-slots.html#Named-Slots" target="_black">official documentation</a> for more information.
+
+    template(v-slot:codes)
+      highlight-code(lang="html" data-title="HTML")
+        pre
+          | &lt;!-- For Vue 2.6.0+ --&gt;
+          |
+          | &lt;!-- Input icon (image) --&gt;
+          | &lt;vue-timepicker&gt;
+          |   &lt;template v-slot:icon&gt;
+          |     &lt;img src="$YOUR_ICON_SRC" /&gt;
+          |   &lt;/template&gt;
+          | &lt;/vue-timepicker&gt;
+          |
+          | &lt;!-- Customized clear button (image) --&gt;
+          | &lt;vue-timepicker&gt;
+          |   &lt;template v-slot:clearButton&gt;
+          |     &lt;img src="$YOUR_CUSTOM_IMAGE_SRC" /&gt;
+          |   &lt;/template&gt;
+          | &lt;/vue-timepicker&gt;
+          |
+          | &lt;!-- Customized dropdown button (character entity) --&gt;
+          | &lt;vue-timepicker manual-input hide-dropdown&gt;
+          |   &lt;template v-slot:dropdownButton&gt;&amp;#x02263;&lt;/template&gt;
+          | &lt;/vue-timepicker&gt;
+
+    template(v-slot:preview)
+      b Input icon (image)
+      p
+        vue-timepicker
+          template(v-slot:icon)
+            img(src="https://i.postimg.cc/CLkZcW46/custom-clock.png")
+      b Customized clear button (image)
+      p
+        vue-timepicker(v-model="customCloseBtnValue")
+          template(v-slot:clearButton)
+            img(src="https://i.postimg.cc/Y0f6RHF8/custom-close.png")
+      b Customized dropdown button (character entity)
+      p
+        vue-timepicker(manual-input hide-dropdown)
+          template(v-slot:dropdownButton) &#x02263;
+
+
+  //- Fixed Dropdown Button
+  sample-block#fixedDropdownButton
+    template(v-slot:title) Fixed Dropdown Button
+    template(v-slot:description)
+      p You can make the dropdown button always visible in the UI with <code>fixed-dropdown-button</code>.
+      p When paired with a customized dropdown button, it's similar to an input icon on the righthand side.
+
+    template(v-slot:codes)
+      highlight-code(lang="html" data-title="HTML")
+        pre
+          | &lt;!-- Default dropdown button --&gt;
+          | &lt;vue-timepicker fixed-dropdown-button&gt;&lt;/vue-timepicker&gt;
+          |
+          | &lt;!-- Customized + fixed dropdown button --&gt;
+          | &lt;vue-timepicker fixed-dropdown-button&gt;
+          |   &lt;template v-slot:dropdownButton&gt;
+          |     &lt;img src="$YOUR_BUTTON_IMAGE_SRC" /&gt; 
+          |   &lt;/template&gt;
+          | &lt;/vue-timepicker&gt;
+          |
+          | &lt;!-- To display only one button on the right --&gt;
+          | &lt;!-- Customized + fixed dropdown button + hide clear button --&gt;
+          | &lt;vue-timepicker fixed-dropdown-button hide-clear-button&gt;
+          |   &lt;template v-slot:dropdownButton&gt;
+          |     &lt;img src="$YOUR_BUTTON_IMAGE_SRC" /&gt; 
+          |   &lt;/template&gt;
+          | &lt;/vue-timepicker&gt;
+
+    template(v-slot:preview)
+      b Default dropdown button
+      p
+        vue-timepicker(fixed-dropdown-button)
+      b Customized + fixed dropdown button
+      p
+        vue-timepicker(fixed-dropdown-button)
+          template(v-slot:dropdownButton)
+            img(src="https://i.postimg.cc/CLkZcW46/custom-clock.png")
+      b Customized + fixed dropdown button + hide clear button
+      p
+        vue-timepicker(fixed-dropdown-button hide-clear-button)
+          template(v-slot:dropdownButton)
+            img(src="https://i.postimg.cc/CLkZcW46/custom-clock.png")
+
+  //- Drop Direction
+  sample-block#dropDirection
+    template(v-slot:title) Drop Direction
+    template(v-slot:description)
+      p Change dropdown direction when needed (<code>v1.1.5+</code>).  Accepting values:
+      ul
+        li
+          b down
+          | : Default value.
+        li
+          b up
+          | : Force open the dropdown above the input.
+        li
+          b auto
+          | : Auto detects available height and opens the dropdown on top if there are not enough spaces below the input.
+
+    template(v-slot:codes)
+      highlight-code(lang="html" data-title="HTML")
+        pre
+          | &lt;!-- Drop Direction Up --&gt;
+          | &lt;vue-timepicker drop-direction="up"&gt;&lt;/vue-timepicker&gt;
+          | &nbsp;
+          | &lt;!-- Auto Drop Direction --&gt;
+          | &lt;div id="auto-dropdown-containter"&gt;
+          |   &lt;!-- Defined Container --&gt;
+          |   &lt;vue-timepicker drop-direction="auto" container-id="auto-dropdown-containter"&gt;&lt;/vue-timepicker&gt;
+          |   &lt;!-- Default (document body) --&gt;
+          |   &lt;vue-timepicker drop-direction="auto"&gt;&lt;/vue-timepicker&gt;
+          | &lt;/div&gt;
+    template(v-slot:preview)
+      b Drop Direction Up
+      p
+        vue-timepicker(drop-direction="up")
+      #auto-dropdown-containter
+        .demo-title
+          b Auto Drop Direction
+        .demo-body
+          .inner
+            b Defined Container
+            p
+              vue-timepicker(drop-direction="auto" container-id="auto-dropdown-containter")
+          .inner
+            b Default (document body)
+            p
+              vue-timepicker(drop-direction="auto")
+
   //- Footer Links
   .footer-links
     slot(name="footer-links")
@@ -889,6 +1032,8 @@ section#mostlyUsedSamples
     width: 20%
     bottom: 0
     left: 0
+    max-height: calc(100vh - 50px)
+    overflow-y: auto
 
     ul
       padding: 0 0 0.5em 1.5em
@@ -908,4 +1053,24 @@ section#mostlyUsedSamples
     height: 2.2em
     padding: 0.3em 0.5em
     font-size: 1em
+  
+  // Drop Direction Demo
+  #auto-dropdown-containter
+    margin-top: 1em
+    padding: 1em
+    background: rgba(0, 0, 0, 0.05)
+    border-radius: 5px
+    .demo-title
+      font-weight: 600
+      padding: 0 0 0.5em 0
+
+    .demo-body
+      display: flex
+      flex-flow: row nowrap
+      justify-content: flex-start
+      align-items: flex-start
+      .inner
+        margin-right: 0.5em
+      b
+        font-weight: normal
 </style>